# docs/swagger.yaml
definitions:
...
  model.Post:
    properties:
      bid:
        type: string
      comments:
        type: integer
      content:
        type: string
      created_at:
        type: string
      creator_id:
        type: string
      img_urls:
        type: string
      likes:
        type: integer
      title:
        type: string
    type: object
  model.PostDraft:
    properties:
      bid:
        type: string
      content:
        type: string
      created_at:
        type: string
      creator_id:
        type: string
      title:
        type: string
    type: object
...
  req.DraftReq:
    properties:
      bid:
        type: string
      sid:
        type: string
    type: object
  req.FindCommentReq:
    properties:
      name:
        type: string
    type: object
  req.LoginReq:
    properties:
      password:
        type: string
      studentid:
        type: string
    type: object
  req.UpdateNameReq:
    properties:
      new_name:
        type: string
      sid:
        type: string
    type: object
  req.UserAvatarReq:
    properties:
      avatar_url:
        type: string
      sid:
        type: string
    type: object
  req.UserSearchReq:
    properties:
      keyword:
        type: string
      sid:
        type: string
    type: object
  resp.LoginResp:
    properties:
      Id:
        type: integer
      avatar:
        type: string
      likes:
        type: string
      name:
        type: string
      school:
        type: string
      sid:
        type: string
      token:
        type: string
    type: object
  resp.Resp:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
  description: 校灵通 API 文档
  title: EventGlide API
paths:
...
  /post/all:
    get:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 获取所有帖子
      tags:
      - Post
  /post/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 帖子
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/model.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 创建帖子
      tags:
      - Post
  /post/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 帖子
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/model.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 删除帖子
      tags:
      - Post
  /post/draft:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 草稿
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/model.PostDraft'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 创建草稿
      tags:
      - Post
  /post/find:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 帖子名
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/req.FindCommentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 通过帖子名查找帖子
      tags:
      - Post
  /post/load:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 草稿请求
        in: body
        name: draft
        required: true
        schema:
          $ref: '#/definitions/req.DraftReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 加载草稿
      tags:
      - Post
  /user/avatar:
    post:
      description: not finished
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户头像更改
        in: body
        name: userAvatarReq
        required: true
        schema:
          $ref: '#/definitions/req.UserAvatarReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 更新头像
      tags:
      - User
  /user/info:
    get:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 学号
        in: query
        name: sid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 获取用户信息
      tags:
      - User
  /user/login:
    post:
      parameters:
      - description: 登录请求
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/req.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  $ref: '#/definitions/resp.LoginResp'
              type: object
      summary: 登录
      tags:
      - User
  /user/logout:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 登出
      tags:
      - User
  /user/search/act:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 搜索请求
        in: body
        name: ureq
        required: true
        schema:
          $ref: '#/definitions/req.UserSearchReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 搜索用户活动
      tags:
      - User
  /user/search/post:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 搜索请求
        in: body
        name: ureq
        required: true
        schema:
          $ref: '#/definitions/req.UserSearchReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 搜索用户帖子
      tags:
      - User
  /user/token/qiniu:
    get:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 获取七牛云token
      tags:
      - User
  /user/username:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新用户名
        in: body
        name: unr
        required: true
        schema:
          $ref: '#/definitions/req.UpdateNameReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 更新用户名
      tags:
      - User
swagger: "2.0"